feat: extract promo code component
